# src/shared/contracts/pipeline-integration.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "VideoPlanet Pipeline Integration API",
    "description": "시나리오→프롬프트→영상 통합 파이프라인 API\n\nBenjamin's Contract-First Architecture:\n- 모든 DTO는 OpenAPI로 정의\n- Consumer-Driven Contract (CDC) 검증\n- 런타임 Zod 스키마 검증\n- 계약 위반 시 컴파일/테스트 실패",
    "version": "1.0.0",
    "contact": {
      "name": "Backend Lead Benjamin",
      "email": "backend@videoplanet.dev"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://videoplanet.vercel.app/api",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/pipeline/status/{projectId}": {
      "get": {
        "summary": "파이프라인 상태 조회",
        "description": "프로젝트의 파이프라인 진행 상태를 조회합니다.",
        "operationId": "getPipelineStatus",
        "tags": ["Pipeline"],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "프로젝트 ID"
          },
          {
            "name": "correlationId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "상관관계 ID (선택적)"
          }
        ],
        "responses": {
          "200": {
            "description": "파이프라인 상태 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineStatusResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "PipelineStatusResponse": {
        "type": "object",
        "required": ["success"],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "required": ["projectId", "currentStep", "status", "progress", "lastUpdated"],
            "properties": {
              "projectId": {
                "type": "string",
                "format": "uuid"
              },
              "currentStep": {
                "type": "string",
                "enum": ["story", "scenario", "prompt", "video"]
              },
              "status": {
                "type": "string",
                "enum": ["idle", "processing", "completed", "failed"]
              },
              "progress": {
                "type": "object",
                "required": ["story", "scenario", "prompt", "video"],
                "properties": {
                  "story": {
                    "type": "object",
                    "required": ["completed"],
                    "properties": {
                      "completed": {
                        "type": "boolean"
                      },
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      }
                    },
                    "additionalProperties": false
                  },
                  "scenario": {
                    "type": "object",
                    "required": ["completed"],
                    "properties": {
                      "completed": {
                        "type": "boolean"
                      },
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      }
                    },
                    "additionalProperties": false
                  },
                  "prompt": {
                    "type": "object",
                    "required": ["completed"],
                    "properties": {
                      "completed": {
                        "type": "boolean"
                      },
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      }
                    },
                    "additionalProperties": false
                  },
                  "video": {
                    "type": "object",
                    "required": ["completed"],
                    "properties": {
                      "completed": {
                        "type": "boolean"
                      },
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "jobId": {
                        "type": "string"
                      }
                    },
                    "additionalProperties": false
                  }
                },
                "additionalProperties": false
              },
              "lastUpdated": {
                "type": "string",
                "format": "date-time"
              },
              "errors": {
                "type": "array",
                "items": {
                  "type": "object",
                  "required": ["step", "message", "timestamp"],
                  "properties": {
                    "step": {
                      "type": "string",
                      "enum": ["story", "scenario", "prompt", "video"]
                    },
                    "message": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "additionalProperties": false
                }
              }
            },
            "additionalProperties": false
          },
          "error": {
            "type": "string"
          }
        },
        "additionalProperties": false
      }
    },
    "responses": {
...
      "NotFoundError": {
        "description": "요청한 리소스를 찾을 수 없음",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "error": {
                  "type": "string",
                  "example": "Story not found with id: 550e8400-e29b-41d4-a716-446655440002"
                }
              }
            }
          }
        }
      },
      "ServerError": {
        "description": "서버 내부 오류",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "error": {
                  "type": "string",
                  "example": "Internal server error"
                },
                "correlationId": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
}